refactor: shared/ui/model, ui 분리
