+++ src/shared/ui/Card/model/type.ts
@@ -1,0 +1,3 @@
+import { ClassNameProps } from "../../type"
+
+export interface CardProps extends React.HTMLAttributes<HTMLDivElement>, ClassNameProps {}
