+++ src/shared/ui/Input/model/type.ts
@@ -1,0 +1,3 @@
+import { ClassNameProps } from "../../type"
+
+export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement>, ClassNameProps {}
