+++ src/shared/ui/Textarea/model/type.ts
@@ -1,0 +1,3 @@
+import { ClassNameProps } from "../../type"
+
+export interface TextareaProps extends React.TextareaHTMLAttributes<HTMLTextAreaElement>, ClassNameProps {}
