+++ src/shared/ui/type.ts
@@ -1,0 +1,3 @@
+export interface ClassNameProps {
+  className?: string
+}
